// frontend/src/config/config.d.ts
export declare const fieldLabels: {
    description: string;
    category: string;
    isActive: string;
    companyBrand: string;
    mediaType: string;
    mediaSubType: string;
    fileSize: string;
    recordedDate: string;
    uploadedBy: string;
    modifiedBy: string;
    metadata: string;
    fileName: string;
    fileExtension: string;
    modifiedDate: string;
    tags: string;
    title: string;
    altText: string;
    productSKU: string;
    visibility: string;
};
export declare const nonEditableFields: {
    _id: boolean;
    id: boolean;
    location: boolean;
    slug: boolean;
    fileExtension: boolean;
    modifiedDate: boolean;
    __t: boolean;
    __v: boolean;
    recordedDate: boolean;
    fileSize: boolean;
    uploadedBy: boolean;
    modifiedBy: boolean;
};
export declare const fieldConfigurations: {
    description: {
        type: string;
        class: string;
    };
    category: {
        type: string;
        options: string[];
    };
    isActive: {
        type: string;
    };
};